Move config methods from FieldGeneratorService to ConfigurationService
