+++ src/services/FieldGeneratorService.php
@@ -23,111 +23,6 @@ class FieldGeneratorService extends Component
     private array $createdBlockEntryTypes = [];
 
     /**
-     * Store configuration data to plugin storage
-     */
-    public function storeConfig(string $name, array $data): string
-    {
-        $plugin = Plugin::getInstance();
-        $plugin->ensureStorageDirectory();
-
-        $configPath = $plugin->getStoragePath() . DIRECTORY_SEPARATOR . 'configs';
-        if (!is_dir($configPath)) {
-            mkdir($configPath, 0755, true);
-        }
-
-        $filename = $this->sanitizeFilename($name) . '_' . time() . '.json';
-        $filepath = $configPath . DIRECTORY_SEPARATOR . $filename;
-
-        if (file_put_contents($filepath, json_encode($data, JSON_PRETTY_PRINT)) === false) {
-            throw new Exception("Failed to write config file: $filepath");
-        }
-
-        $this->cleanupOldConfigs($configPath);
-
-        return $filepath;
-    }
-
-    /**
-     * Get stored configuration
-     */
-    public function getStoredConfig(string $filename): ?array
-    {
-        $plugin = Plugin::getInstance();
-        $configPath = $plugin->getStoragePath() . DIRECTORY_SEPARATOR . 'configs' . DIRECTORY_SEPARATOR . $filename;
-
-        if (!file_exists($configPath)) {
-            return null;
-        }
-
-        $data = json_decode(file_get_contents($configPath), true);
-        return $data ?: null;
-    }
-
-    /**
-     * List all stored configurations
-     */
-    public function listStoredConfigs(): array
-    {
-        $plugin = Plugin::getInstance();
-        $configPath = $plugin->getStoragePath() . DIRECTORY_SEPARATOR . 'configs';
-
-        if (!is_dir($configPath)) {
-            return [];
-        }
-
-        $configs = [];
-        $files = glob($configPath . DIRECTORY_SEPARATOR . '*.json');
-
-        foreach ($files as $file) {
-            $configs[] = [
-                'filename' => basename($file),
-                'path' => $file,
-                'created' => filemtime($file),
-                'size' => filesize($file),
-            ];
-        }
-
-        // Sort by creation time, newest first
-        usort($configs, fn($a, $b) => $b['created'] <=> $a['created']);
-
-        return $configs;
-    }
-
-    /**
-     * Clean up old configuration files
-     */
-    private function cleanupOldConfigs(string $configPath): void
-    {
-        $settings = Plugin::getInstance()->getSettings();
-        $maxConfigs = $settings->maxStoredConfigs;
-
-        $files = glob($configPath . DIRECTORY_SEPARATOR . '*.json');
-        if (count($files) <= $maxConfigs) {
-            return;
-        }
-
-        // Sort by modification time
-        usort($files, fn($a, $b) => filemtime($a) <=> filemtime($b));
-
-        // Delete oldest files
-        $filesToDelete = array_slice($files, 0, count($files) - $maxConfigs);
-        foreach ($filesToDelete as $file) {
-            unlink($file);
-        }
-    }
-
-    /**
-     * Sanitize filename
-     */
-    private function sanitizeFilename(string $name): string
-    {
-        // Remove or replace problematic characters
-        $name = preg_replace('/[^a-zA-Z0-9_-]/', '_', $name);
-        $name = trim($name, '_');
-        return $name ?: 'config';
-    }
-
-    /**
      * Create a field from config array
      * This method handles the normalized config format from LLM responses
      */
